made some changes to the shell script, tested ebs auto mounting with ec2 instances from scratch

+++ ebs-available/shell/scripts_for_mounting_ebs.sh
@@ -5,33 +5,33 @@ logic=true
 value="data"
 n=0
 
-file -s $device # Rajesh has this part of the script. There should be an if statement here
-mkfs -t xfs $device
+# first volume is taken care by Amazon
+if [ "$device" != "/dev/xvda" ]; then # this would avoid going for the first volume that's created
+    file -s $device
+    mkfs -t xfs $device
 
-while [ "$logic" != "false" ]
-do
-    x=`ls / | grep $value`
-    # mkdir
-    if [[ -z  $x ]]; then
-        mkdir /$value
-        logic=false
-    else
-        ((n+=1)) #sum operations
-        value="data"
-        value="$value$n" # assigning new value
-    fi
-done
-
-mount $device /$value
-cp /etc/fstab /etc/fstab.orig
-
-# there is another piece of logic to be done here
-# have to strip the shit out of this
-blkid=`blkid -o value -s UUID $device` # Rajesh also has this part of the script
-# would XFS have to be constant or a variable?
-# would the number 2 have to auto increase?
-UUID=`$blkid  /$value  xfs  defaults,nofail  0  2`
-cat $UUID >> fstab
-
-umount /$value
-mount -a
+    while [ "$logic" != "false" ]
+    do
+        x=`ls / | grep $value`
+        # mkdir
+        if [[ -z  $x ]]; then
+            mkdir /$value
+            logic=false
+        else
+            ((n+=1)) #sum operations
+            value="data"
+            value="$value$n" # assigning new value
+        fi
+    done
+    
+    echo $device /$value >> /home/ec2-user/test.txt
+    mount $device /$value
+    
+    cp /etc/fstab /etc/fstab.orig
+    blkid=`blkid -o value -s UUID $device`
+    echo $blkid >> /home/ec2-user/test.txt
+    UUID=`$blkid  /$value  xfs  defaults,nofail  0  2`
+    echo $UUID >> /etc/fstab
+    
+    echo >> /home/ec2-user/test.txt
+fi
